requirements.txt
flask==3.0.2
openai==1.12.0
whisper==1.1.10
moviepy==1.0.3
requests==2.31.0
Werkzeug==3.0.1
uuid==1.30
python-dotenv==1.0.1
gunicorn==21.2.0
